Made changes to methods so they work with tables in general instead of manipulating this.table directly. This is because 2 tables will exist when the simulation is running: the previous and the next generation.
Tables are now stored in an arra: this.tables = []
At most, this array will have 2 tables and the lowest index table will be
the older one.

I added methods to turn on/off the toggle state event listener of cells.
This is because there might be situations where it should be turned off. For example,
if the new table is already calculated but there is an animation going that will eventually
display it but it still displays the old table, clicking on the cells at that moment will not
impact the new calculation (since it was already calculated but it hasn't been displayed).

 Changes to be committed:
	modified:   gameoflife.js
	modified:   styles.css